actor: Add ClutterActorClass.layout_children()
The allocate() virtual is far too complex, and it has far too many
things that can possibly go wrong:

  - it needs to be chained up, to store the allocation of the actor;
    - unless you call set_allocation() to do that;
  - you can override it;
    - but you can also have a layout manager delegate;
  - you have to allocate() all the children after you chained up;
  - you may get signals mid-way through the allocation process;
  - you get a non-normalized box;
    - unless you're using a layout manager, which will get a
      normalized one.

Let's try to keep this stuff simple, and minimize the amount of things
to remember to get it right.

First of all, we add a new virtual, called layout_children(); it gets
passed the actor itself, and it will be called only if the actor has
children to lay out. Its default implementation will call the layout
manager's allocation method, if one is set. Inside the layout_children()
virtual, you will need to iterate over the children, compute their
bounds and call clutter_actor_allocate(), similarly to how you implement
a paint() virtual. The layout_children() virtual is called after the
allocation of the actor has been stored, so it's possible to safely
query it without hitting slow code paths. You do not need to chain up,
and you do not need to normalize boxes. All the signals and property
notifications on the parent will happen before the layout_children()
virtual has been called, so the actor's state is stable by the time you
call allocate() on the children.