stage: Add a paint callback
The ClutterActor::paint signal is deprecated, and connecting to it even
to get notifications will disable clipped redraws because of violations
of the paint volume.

The only actual valid use case for notifications of a successful frame
is on the ClutterStage, so we should add new (experimental) API for it,
so that users can actually subscribe to it — at least if you're writing
a compositor.

Shoving a signal in a performance critical path is not an option, and
I'm not sure I want to commit to an API like this yet. I reserve the
right to revisit this decision in the future.

This is a backport of the same commit from the clutter-1.16 branch, and
it introduces new API in a stable series.

https://bugzilla.gnome.org/show_bug.cgi?id=698783

(cherry picked from commit d061a47573fbfec69ed6f2fd02f233e218830a6d)

Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>