actor: Use the real opacity when clearing the stage
The paint opacity for a top level is always overridden to be the full
value, since it's a composited value and we want to paint our scene.

When clearing the stage framebuffer, though, we want to use the actual
opacity, if ClutterStage:use-alpha is set.

(cherry picked from commit 8b8aa0b4927bc865b08b77b4e28237da0ff8597c)

Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>