build: Reset interface_age
We don't usually add symbols inside a stable branch, but given that the
newly added clutter_gdk_get_visual() function is necessary to make
Clutter work when embedded inside GTK running on nVidia binary drivers,
I decided to break the rule. This requires resetting the interface age,
so that the soname version changes.

https://bugzilla.gnome.org/show_bug.cgi?id=747489